refactor: make gateway startup hook self-contained, drop direct tool
- Remove tools/agentlair_tool.py — tool is provided by the hermes-agentlair
  plugin (on_session_start / send_agentlair_message). Having two tool
  registrations for the same name causes a conflict.

- Rewrite gateway/builtin_hooks/agentlair_inbox_drain.py to be self-contained
  using stdlib urllib only — no dependency on the removed tool file or on
  the plugin package being installed. Switches env var to AGENTLAIR_ADDRESS
  (consistent with the plugin's convention).

- Revert model_tools.py addition (tool is plugin-provided, not core).

The builtin hook still fires once at gateway:startup for initial inbox drain,
while the hermes-agentlair plugin handles per-session drain via on_session_start.
