+++ model_tools.py
@@ -158,7 +158,6 @@ def _discover_tools():
         "tools.send_message_tool",
         # "tools.honcho_tools",  # Removed — Honcho is now a memory provider plugin
         "tools.homeassistant_tool",
-        "tools.agentlair_tool",
     ]
     import importlib
     for mod_name in _modules:
